Factor out county wizard audit board sign-in.
