+++ client/src/component/county/audit/wizard/CountyAuditWizard.tsx
@@ -2,44 +2,9 @@ import * as React from 'react';
 
 import { Checkbox, EditableText, Radio, RadioGroup } from '@blueprintjs/core';
 
+import AuditBoardSignIn from './AuditBoardSignIn';
 import WizardStart from './WizardStart';
 
-
-const AuditBoardSignInForm = () => (
-    <div>
-        <h3>Audit Board Member</h3>
-        <div className='pt-card'>
-            <label>Full Name:
-                <EditableText />
-            </label>
-        </div>
-        <div className='pt-card'>
-            <RadioGroup label='Party Affiliation' onChange={ () => ({}) }>
-                <Radio label='Democratic Party' />
-                <Radio label='Republican Party' />
-                <Radio label='Minor Party' />
-                <Radio label='Unaffiliated' />
-            </RadioGroup>
-        </div>
-    </div>
-);
-
-const AuditBoardSignIn = ({ nextStage }: any) => (
-    <div>
-        <div>
-            <h2>Audit Board Sign-in</h2>
-            <p>Enter the full names and party affiliations of each member of
-                the Acme County Audit Board who will be conducting this audit
-                today:
-            </p>
-        </div>
-        <AuditBoardSignInForm />
-        <AuditBoardSignInForm />
-        <button className='pt-button pt-intent-primary' onClick={ nextStage }>
-            Next
-        </button>
-    </div>
-);
 
 const AuditInstructions = () => (
     <div>
